#20 refactor: SettingTapTableViewCell, SettingTextTapTableViewCell을 SettingTapActionTableViewCell로 통합
- Tap Action을 나타내는 expand_right 이미지와 additionalTextLabel을 기본으로 갖고, additionalTextLabel의 text는 옵셔널로 나타냄
